+++ ReSwiftTests/AutomaticallySkipRepeatsTests.swift
@@ -27,20 +27,39 @@ class AutomaticallySkipRepeatsTests: XCTestCase {
         super.tearDown()
     }
 
-    func testInitialSubscription() {
+    func testInitialSubscriptionWithRegularSubstateSelection() {
         store.subscribe(self) { $0.select { $0.name } }
         XCTAssertEqual(self.subscriptionUpdates, 1)
     }
 
-    func testDispatchUnrelatedActionWithExplicitSkipRepeats() {
+    func testInitialSubscriptionWithKeyPath() {
+        store.subscribe(self) { $0.select(\.name) }
+        XCTAssertEqual(self.subscriptionUpdates, 1)
+    }
+
+    func testDispatchUnrelatedActionWithExplicitSkipRepeatsWithRegularSubstateSelection() {
         store.subscribe(self) { $0.select { $0.name }.skipRepeats() }
         XCTAssertEqual(self.subscriptionUpdates, 1)
         store.dispatch(ChangeAge(newAge: 30))
         XCTAssertEqual(self.subscriptionUpdates, 1)
     }
 
-    func testDispatchUnrelatedActionWithoutExplicitSkipRepeats() {
+    func testDispatchUnrelatedActionWithExplicitSkipRepeatsWithKeyPath() {
+        store.subscribe(self) { $0.select(\.name).skipRepeats() }
+        XCTAssertEqual(self.subscriptionUpdates, 1)
+        store.dispatch(ChangeAge(newAge: 30))
+        XCTAssertEqual(self.subscriptionUpdates, 1)
+    }
+
+    func testDispatchUnrelatedActionWithoutExplicitSkipRepeatsWithRegularSubstateSelection() {
         store.subscribe(self) { $0.select { $0.name } }
+        XCTAssertEqual(self.subscriptionUpdates, 1)
+        store.dispatch(ChangeAge(newAge: 30))
+        XCTAssertEqual(self.subscriptionUpdates, 1)
+    }
+
+    func testDispatchUnrelatedActionWithoutExplicitSkipRepeatsWithKeyPath() {
+        store.subscribe(self) { $0.select(\.name) }
         XCTAssertEqual(self.subscriptionUpdates, 1)
         store.dispatch(ChangeAge(newAge: 30))
         XCTAssertEqual(self.subscriptionUpdates, 1)