+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Drop support for Swift 3.2, 4.0, and 4.1. (#418) - @DivineDominion
 
 **Other:**
+- Add key paths to subscription select (#415) - @djtech42
 - Make isDispatching of Store atomic (#341) - @zhongwuzw
 
 # 5.0.0